Move BindUpdateConstraints into a virtual function that is implemented by the DuckTableEntry
